Replaced manual CohtmlDataPatch with harmony annotation, fixed UIX causing ex8 menu to not load, adjusted search button to not be offset. Fixed categories not being alphabetical
