+++ UIExpansionKit/WebUi/Js/main-menu-inject.js
@@ -227,3 +227,9 @@ UIX_updateCategoriesBodies()
 
 let settingsButton = document.getElementsByClassName("tab_btn_settings")[0]
 settingsButton.insertAdjacentHTML('afterend', `<div class="toolbar-btn tab_btn_settings button" onclick="changeTab(\'uix-settings\', this);"><img src="uix-resource:twemoji-melon.svg"/>Mòwóds</div>`)
+
+let searchButton = document.getElementById("search-btn")
+searchButton.style.width = "95%"
+searchButton.style.paddingLeft = "8px"
+searchButton.style.paddingTop = "5px"
+
